Act CI/CD 4

+++ scripts/application_stop.sh
@@ -1,4 +1,9 @@
 #!/bin/bash
 #Stopping existing node servers
 echo "Deteniendo servicios desplegados de ReactJs"
-sudo kill -9 $(sudo lsof -t -i:3000)
+
+if [ sudo lsof -t -i:3000 ]; then
+  sudo kill -9 $(sudo lsof -t -i:3000) 
+else
+  echo "Not working"
+fi
